+++ lib/rubocop/cop/style/multiline_method_call_brace_layout.rb
@@ -84,6 +84,15 @@ module RuboCop
 
           args
         end
+
+        def ignored_literal?(node)
+          single_line_ignoring_receiver?(node) || super
+        end
+
+        def single_line_ignoring_receiver?(node)
+          return false unless node.loc.begin && node.loc.end
+          node.loc.begin.line == node.loc.end.line
+        end
       end
     end
   end