+++ spec/rubocop/cop/style/multiline_method_call_brace_layout_spec.rb
@@ -56,5 +56,18 @@ describe RuboCop::Cop::Style::MultilineMethodCallBraceLayout, :config do
       inspect_source(cop, 'puts("Hello world!")')
       expect(cop.offenses).to be_empty
     end
+
+    it 'ignores single-line calls with multi-line receiver' do
+      inspect_source(cop, ['[',
+                           '].join(" ")'])
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'ignores single-line calls with multi-line receiver with leading dot' do
+      inspect_source(cop, ['[',
+                           ']',
+                           '.join(" ")'])
+      expect(cop.offenses).to be_empty
+    end
   end
 end